+++ env-var.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const variable = require('./lib/variable')
+const EnvVarError = require('./lib/env-error')
 
 /**
  * Returns an "env-var" instance that reads from the given container of values.
@@ -23,15 +24,18 @@ const from = (container, extraAccessors) => {
     /**
      * Returns a variable instance with helper functions, or process.env
      * @param  {String} variableName Name of the environment variable requested
-     * @param  {String} defaultValue Optional default to use as the value
      * @return {Object}
      */
-    get: (variableName, defaultValue) => {
+    get: function (variableName) {
       if (!variableName) {
         return container
       }
 
-      return variable(container, variableName, defaultValue, extraAccessors || {})
+      if (arguments.length > 1) {
+        throw new EnvVarError('It looks like you passed more than one argument to env.get(). Since env-var@6.0.0 this is no longer supported. To set a default value use env.get(TARGET).default(DEFAULT)')
+      }
+
+      return variable(container, variableName, extraAccessors || {})
     }
   }
 }